Item quantity is now double
Changed Item and OrderItem quantity from int to double
Remove order button is disabled without selection
Moved remove item button to stock screen
Moved Stock screen buttons to the side
Combined new and edit item screens
Split the DBBroker into 3 pieces
Centered item screen + other ui improvements

+++ src/model/containers/Item.java
@@ -10,7 +10,7 @@ public class Item implements Serializable
 	 */
 	private static final long serialVersionUID = 1L;
 	private double id;
-	private int quantity;
+	private double quantity;
 	private String name;
 
 //===================== Constructor ===========================================
@@ -22,7 +22,7 @@ public class Item implements Serializable
 		this.name = "This Item does not exist";
 	}
 	
-	public Item(double id, int quantity, String name)
+	public Item(double id, double quantity, String name)
 	{
 		this.id = id;
 		this.name = name;		
@@ -51,11 +51,11 @@ public class Item implements Serializable
 		this.id = id;
 	}
 
-	public int getQuantity() {
+	public double getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(int quantity) {
+	public void setQuantity(double quantity) {
 		this.quantity = quantity;
 	}
